requirements.txt
appdirs==1.4.3
attrs==19.1.0
bcrypt==3.1.7
black==19.3b0
blinker==1.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Caching==1.7.2
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.1.0
pycparser==2.19
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
toml==0.10.0
urllib3==1.25.5
Werkzeug==0.15.6
WTForms==2.2.1
